optional dependecies , remaining nix and cty
