+++ Cargo.toml
@@ -3,37 +3,56 @@ name = "growbe-mainboard"
 version = "0.1.10"
 edition = "2021"
 
+[features]
+default = ["com_virtual", "http_server", "com_ws", "com_ble", "com_i2c", "com_virt"]
+gpio = ["rppal"]
+com_ble = ["bluer", "btleplug", "futures", "uuid", "tokio-stream"]
+com_virtual = []
+com_ws = ["tokio-tungstenite", "futures", "futures-util", "url"]
+com_i2c = ["gpio"]
+com_virt = []
+http_server = ["warp"]
+
 [dependencies]
 chrono = "0.4.0"
 rumqtt = "0.31.0"
-cty = "0.2.2"
 env_logger = "0.9.0"
 log = "0.4.14"
 protobuf = "2.25.2"
-uuid = "1.1.2"
-futures = "0.3"
 lazy_static = "1.4.0"
 rusqlite = { version = "0.28.0", features = ["bundled"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.59"
 tokio = { version = "1.14", features = ["full", "time", "rt-multi-thread"] }
 tokio-util = "0.6.9"
-tokio-stream = { version = "0.1.9", features = ["sync"] }
-warp = "0.3"
-nix = "0.23.0"
-
-tokio-tungstenite = "*"
-futures-util = { version = "0.3", default-features = false, features = ["sink", "std"] }
-url = "2.0.0"
-
 regex = "1.5"
 
+nix = { version = "0.23.0", optional = false }
+
+cty = { version = "0.2.2", optional = false }
+
+futures = { version = "0.3", optional = true }
+
+# dependencies for web server
+warp = { version = "0.3" , optional = true}
+
+# dependencies for web socket
+tokio-tungstenite = { version = "*",  optional = true }
+futures-util = { version = "0.3", default-features = false, features = ["sink", "std"], optional = true }
+url = { version = "2.0.0", optional = true }
+tokio-stream = { version = "0.1.9", features = ["sync"], optional = true }
+
+# dependencies for ble
+uuid = { version = "1.1.2", optional = true }
+
 [target.'cfg(any(target_os = "macos", target_os = "windows"))'.dependencies]
-btleplug = { version = "0.10", features = ["serde"] }
+btleplug = { version = "0.10", features = ["serde"], optional = true }
 
 [target.'cfg(target_os = "linux")'.dependencies]
-rppal = "0.13.1"
-bluer = {version = "0.15.0", features = ["bluetoothd"]}
+bluer = {version = "0.15.0", features = ["bluetoothd"], optional = true}
+# for gpio
+rppal = { version = "0.13.1", optional = true }
+
 
 [build-dependencies]
 protoc-rust = "2.25.2"
