+++ src/comboard/imple/mod.rs
@@ -1,7 +1,9 @@
 pub mod interface;
-#[cfg(target_os = "linux")]
+pub mod channel;
+#[cfg(all(target_os = "linux", feature = "com_i2c"))]
 pub mod i2c_linux;
+#[cfg(feature = "com_ws")]
 pub mod ws;
 pub mod virt;
+#[cfg(feature = "com_ble")]
 pub mod ble;
-pub mod channel;
