+++ src/main.rs
@@ -9,13 +9,12 @@ mod id;
 mod mainboardstate;
 mod modulestate;
 mod store;
-mod server;
 mod plateform;
 mod utils;
+mod server;
 
 use std::sync::{Mutex, Arc, mpsc::channel};
 
-use tokio::select;
 
 #[tokio::main(flavor = "multi_thread")]
 async fn main() {
@@ -65,25 +64,11 @@ async fn main() {
 
 
 
+    #[cfg(feature = "http_server")]
     let server_task = server::http::get_server(&crate::mainboardstate::config::CONFIG.server);
+    #[cfg(not(feature = "http_server"))]
+    let server_task  = async {};
 
-    // Wait for all task to finish (they should never end)
-    //let _ret = tokio::join!(
-    //    server_task,
-    //    module_state_task,
-    //    socket_task,
-    //);
 
-    select! {
-        _ = server_task => {},
-        _ = module_state_task => {},
-        _ = socket_task => {},
-    }
-
-    //let exit_code = if let Err(_) = ret {
-    //    1
-    //} else {
-    //    0
-    //};
-    //std::process::exit(exit_code);
+    let _ = tokio::join!(server_task, module_state_task, socket_task);
 }
