+++ src/server/mod.rs
@@ -1,3 +1,5 @@
 
+pub mod config;
+#[cfg(feature = "http_server")]
 pub mod http;
